requirements.txt
Flask==2.0.1
h5py==3.1.0
Jinja2==3.0.1
gevent==21.1.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
numpy==1.19.5
Pillow==8.2.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
Werkzeug==2.0.1
gunicorn==20.1.0